\name{betaEst}
\alias{betaEst}
\title{Beta Distribution Maximum Likelihood Estimator}
\description{
Estimates a beta distribution via Maximum Likelihood
}
\usage{
betaEst(y, w, weights)
}
\arguments{
  \item{y}{data vector}
  \item{w}{posterior weights}
  \item{weights}{case weights}
}
\details{
  Typically not be called by user.
}
\value{
(a,b) parameters
}
\keyword{cluster}
