\name{blcTreeLeafClasses}
\alias{blcTreeLeafClasses}
\title{Posterior Class Assignments for Beta RPMM}
\description{
Gets a vector of posterior class membership assignments for terminal nodes.
}
\usage{
blcTreeLeafClasses(tr)
}
\arguments{
  \item{tr}{Tree from which to create assignments.}
}
\details{
  See \code{\link{blcTree}} for example.
}
\value{
  Vector of class assignments
}
\seealso{\code{\link{blcTreeLeafMatrix}} }
\keyword{tree}
\keyword{cluster}

