\name{glcTreeLeafClasses}
\alias{glcTreeLeafClasses}
\title{Posterior Class Assignments for Gaussian RPMM}
\description{
Gets a vector of posterior class membership assignments for terminal nodes.
}
\usage{
glcTreeLeafClasses(tr)
}
\arguments{
  \item{tr}{Tree from which to create assignments.}
}
\details{
  See \code{\link{glcTree}} for example.
}
\value{
  Vector of class assignments
}
\seealso{\code{\link{glcTreeLeafMatrix}} }
\keyword{tree}
\keyword{cluster}

