\name{IlluminaMethylation}
\docType{data}
\alias{IlluminaMethylation}
\alias{tissue}
\alias{IllumBeta}
\title{DNA Methylation Data for Normal Tissue Types}
\description{
Illumina GoldenGate DNA methylation data for 217 normal tissues.
100 most variable CpG sites.
}
\usage{IlluminaMethylation}
\format{a 217 x 100 matrix containing Illumina Avg Beta values (IllumBeta), 
and a corresponding factor vector of 217 tissue types (tissue).}
\references{
Christensen BC, Houseman EA, et al. 2009 Aging and Environmental Exposures Alter Tissue-Specific DNA Methylation Dependent upon CpG Island Context. PLoS Genet 5(8): e1000602.
}
\keyword{datasets}