\name{plot.blcTree}
\alias{plot.blcTree}
\title{Plot a Beta RPMM Tree Profile}
\description{
Plot method for objects of type \dQuote{blcTree}.  Plots profiles of terminal nodes in color.
Method wrapper for \code{plotImage.blcTree}.
}
\usage{
\method{plot}{blcTree}(x,...)
}
\arguments{
  \item{x}{RPMM object to plot.}
  \item{\dots}{Additional arguments to pass to \code{plotImage.blcTree}.}
}
\details{
   See \code{\link{blcTree}} for example.
}
\keyword{tree}
\keyword{cluster}
