\name{plot.glcTree}
\alias{plot.glcTree}
\title{Plot a Gaussian RPMM Tree Profile}
\description{
Plot method for objects of type \dQuote{glcTree}.  Plots profiles of terminal nodes in color.
Method wrapper for \code{plotImage.glcTree}.
}
\usage{
\method{plot}{glcTree}(x,...)
}
\arguments{
  \item{x}{RPMM object to plot.}
  \item{\dots}{Additional arguments to pass to \code{plotImage.glcTree}.}
}
\details{
   See \code{\link{glcTree}} for example.
}
\keyword{tree}
\keyword{cluster}
