\name{print.blcTree}
\alias{print.blcTree}
\title{Print a Beta RPMM object}
\description{
Print method for objects of type blcTree
}
\usage{
\method{print}{blcTree}(x,...)
}
\arguments{
  \item{x}{RPMM object to print}
  \item{\dots}{(Unused).}
}
\details{
  See \code{\link{blcTree}} for example.
}
\keyword{tree}
\keyword{cluster}
