\name{print.glcTree}
\alias{print.glcTree}
\title{Print a Gaussian RPMM object}
\description{
Print method for objects of type blcTree
}
\usage{
\method{print}{glcTree}(x,...)
}
\arguments{
  \item{x}{RPMM object to print}
  \item{\dots}{(Unused).}
}
\details{
  See \code{\link{glcTree}} for example.
}
\keyword{tree}
\keyword{cluster}

